+++ streams/src/at/dms/kjc/sir/lowering/partition/cache/CConfigPipeline.java
@@ -414,8 +414,8 @@ class CConfigPipeline extends CConfigContainer {
         // check if we have precomputed
         if (fusion_info[from][to] != null) return fusion_info[from][to];
 
-        int work = 0;
-        int work_no_penalty = 0;
+        long work = 0;
+        long work_no_penalty = 0;
         int code = 0;
         int data = 0;
